Added driver controls and added ball sensor for the intake
I added the drive controls in the MyRobot.cpp file. They are as follows:
R-Trigger: Shoot Catapult
R-2: Move the intake in
R-3: Move in intake only enough to pass
R-5: Move the intake out
R-6 & L-10: Stop the catapult

Also, yesterday I added a sensor that will determine if the ball is in
the intake for if we want to pass the ball.

Bijan