Added waits to Catapult and SmartDashboard functionality
I put a wait of 0.9 seconds in all places after the catapult shoots. The
purpose of this is to allow the catapult to finish shooting before
pulling it back.

I also added the ability for the driver to type in the type of
autonomous we are going to use in a match using SmartDashboard.

Bijan